+++ data/resources/CHANGELOG.md
@@ -1,5 +1,6 @@
 Jun 26, 2023:
 - Fix CTextureAnimation (#729). This should fix ImGui drawing of game icons
+- lua: Disallow the use of mq.delay while a module is being imported (#730)
 
 June 21, 2023:
 - live: Updated for live patch